+++ sdk/core/azure-core/azure/core/pipeline/transport/requests_basic.py
@@ -120,14 +120,14 @@ class StreamDownloadGenerator(object):
                     retry_active = False
                 else:
                     time.sleep(retry_interval)
-                    headers = {'range': 'bytes=' + self.downloaded + '-'}
+                    headers = {'range': 'bytes=' + str(self.downloaded) + '-'}
                     resp = self.pipeline.run(self.request, stream=True, headers=headers)
                     if resp.status_code == 416:
                         raise
                     chunk = next(self.iter_content_func)
                     if not chunk:
                         raise StopIteration()
-                    self.downloaded += chunk
+                    self.downloaded += len(chunk)
                     return chunk
                 continue
             except requests.exceptions.StreamConsumedError: